requirements.txt
biopython>=1.79
pandas>=1.4
numpy>=1.21
tqdm>=4.64

